Update package.json
Updated the username for my internal CI to download from this branch. May be the git might have a special command for this need but I don't research before doing this. I was in a hurry.
Now reversed to original so that this this request can be merge with master quickly.