Support extraOffsets
React Native Charts Wrapper now supports the extraOffsets feature for
BarChart, HorizontalBarChart, and LineChart, on iOS and Android.

The extraOffsets property allows developers to overcome clipping issues
such as http://github.com/PhilJay/MPAndroidChart/issues/1657 or
http://stackoverflow.com/q/32169593/2619107.

Bug fix in the setViewPortOffsets method of "RNBarLineChartViewBase. It
was using the left JSON property for the top parameter of the underlying
barLineChart.setViewPortOffsets call.